# Suggested changes to speed up phaser