@vincentqb 2020-05-26
> * How does this behave on GPU?

Let me know how the update behaves :)