@cpuhrsch 2020-05-28
@vincentqb - I'd actually prefer to hand this off. There is likely a much bigger performance gain to be had by unrolling parts of the for-loop (delay_buf_len steps) and vectorizing that part.