@cpuhrsch 2020-05-28
cc @bhargavkathivarapu as the original author of phaser, I think this could be interesting for you